+++ src/HyperFormula.ts
@@ -2415,7 +2415,6 @@ export class HyperFormula implements TypedEmitter {
    * @param {SimpleCellRange} target range where data is intended to be put
    * @param {boolean} offsetsFromTarget if true, offsets are computed from target corner, otherwise from source corner
    *
-   * @throws [[SheetsNotEqual]] if both ranges are not from the same sheet
    * @throws [[EvaluationSuspendedError]] when the evaluation is suspended
    * @throws [[ExpectedValueOfTypeError]] if source or target are of wrong type
    * @throws [[SheetsNotEqual]] if range provided has distinct sheet numbers for start and end
@@ -2440,16 +2439,13 @@ export class HyperFormula implements TypedEmitter {
     }
     const sourceRange = new AbsoluteCellRange(source.start, source.end)
     const targetRange = new AbsoluteCellRange(target.start, target.end)
-    if (sourceRange.sheet !== targetRange.sheet) {
-      throw new SheetsNotEqual(sourceRange.sheet, targetRange.sheet)
-    }
     this.ensureEvaluationIsNotSuspended()
     return targetRange.arrayOfAddressesInRange().map(
       (subarray) => subarray.map(
         (address) => {
           const row = ((address.row - (offsetsFromTarget ? target : source).start.row) % sourceRange.height() + sourceRange.height()) % sourceRange.height() + source.start.row
           const col = ((address.col - (offsetsFromTarget ? target : source).start.col) % sourceRange.width() + sourceRange.width()) % sourceRange.width() + source.start.col
-          return this._serialization.getCellSerialized({row, col, sheet: targetRange.sheet}, address)
+          return this._serialization.getCellSerialized({row, col, sheet: sourceRange.sheet}, address)
         }
       )
     )